C# code
This code uses an example array of numbers. We create two int variables to which we assign the values of the first item in the array. After that we run a foreach loop, running through each element in the array comparing the values and thus, finding the smallest and largest number in the array.